Use `ll` instead of `int` everywhere (#4)
* Change all `int`s by `ll`
* Increase size of buf in `BumpAllocator` so that there is enough memory for HLD test
* Use bigger `inf`s

+++ stress-tests/number-theory/ModLog.cpp
@@ -3,12 +3,12 @@
 #include "../../content/number-theory/ModLog.h"
 
 int main() {
-	const int lim = 100;
+	const ll lim = 100;
 	rep(m,1,lim) {
 		rep(a,0,lim) {
 			vector<ll> ans(m, -1);
 			ll b = a % m;
-			rep(x,1,max(m,2)) {
+			rep(x,1,max(m,2ll)) {
 				if (ans[b] == -1) ans[b] = x;
 				b = b * a % m;
 			}
